runtime-rs: refactor runtime handler
In the previous design, runtime instance is used for different
runtime wasm, linux and virt. Sandbox and container manager trait are
defined inside it, most of the time we just use one of them for any
operation. In some case, like update sandbox resource, we need to call
both of them, it's hard to do based on the previous design. It's better to
provide a higher level abstraction on rutime handler, move sandbox and
container manager in the runtime struct itself.

Signed-off-by: Zhongtao Hu <zhongtaohu.tim@linux.alibaba.com>