dragonball: change the storage location of sev_start and sev_secret
Wrap `sev_start` with `Box`, moving it from the stack to the heap, as it
is a 2KB data structure.

Move `sev_secret` from `VmConfigInfo` to `Vm`, because it's not a pre-boot
configuration item, but provided in the middle of the two-stage boot
process; and it also moves from the stack to the heap.

Signed-off-by: Fuu <fuu-open@linux.alibaba.com>